add vendor version check script
this may or may not be kanged from jesec.
Also you'd be surprised by how many times some random peep kept insisting he was on the right vendor and it was the rom's fault it was bootlooping and when i asked for an vendor/build.prop it turns out it was some oldass oreo shit.